fix JVM tests (which are JVM main actually)
