+++ shared/src/jvmMain/kotlin/Main.kt
@@ -1,12 +1,43 @@
 package dev.johnoreilly.confetti
 
+import com.apollographql.apollo3.ApolloClient
+import com.apollographql.apollo3.api.ApolloResponse
+import com.apollographql.apollo3.api.DefaultFakeResolver
+import com.apollographql.apollo3.api.FakeResolverContext
+import com.apollographql.apollo3.testing.MapTestNetworkTransport
+import com.benasher44.uuid.uuid4
 import dev.johnoreilly.confetti.di.initKoin
+import dev.johnoreilly.confetti.schema.__Schema
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withTimeout
+import kotlinx.datetime.LocalDateTime
+import org.koin.dsl.module
+
+private fun testModule() = module {
+    factory {
+        ApolloClient.Builder()
+            .networkTransport(MapTestNetworkTransport().apply {
+                register(GetSessionsQuery(), ApolloResponse.Builder(
+                    GetSessionsQuery(),
+                    uuid4(),
+                    GetSessionsQuery.Data(object: DefaultFakeResolver(__Schema.all) {
+                        override fun resolveLeaf(context: FakeResolverContext): Any {
+                            return when(context.mergedField.type.rawType().name) {
+                                "LocalDateTime" -> return LocalDateTime(1970, 1, 1, 1, 1, 1)
+                                else -> super.resolveLeaf(context)
+                            }
+                        }
+                    }) {}
+                ).build())
+            })
+    }
+}
 
 @Suppress("UNUSED_PARAMETER")
 suspend fun main(args: Array<String>) {
-    val koin = initKoin().koin
+    val koin = initKoin {
+        modules(testModule())
+    }.koin
     val repo = koin.get<ConfettiRepository>()
     val clientCache = koin.get<ApolloClientCache>()
